Remove db2.py and move the code to db.py
